Added Controller for twitter apis
