+++ main.py
@@ -33,6 +33,7 @@ from superagi.controllers.budget import router as budget_router
 from superagi.controllers.config import router as config_router
 from superagi.controllers.organisation import router as organisation_router
 from superagi.controllers.project import router as project_router
+from superagi.controllers.twitter_oauth import router as twitter_oauth_router
 from superagi.controllers.resources import router as resources_router
 from superagi.controllers.tool import router as tool_router
 from superagi.controllers.tool_config import router as tool_config_router
@@ -104,6 +105,7 @@ app.include_router(tool_config_router, prefix="/tool_configs")
 app.include_router(config_router, prefix="/configs")
 app.include_router(agent_template_router, prefix="/agent_templates")
 app.include_router(agent_workflow_router, prefix="/agent_workflows")
+app.include_router(twitter_oauth_router, prefix="/twitter")
 
 
 # in production you can use Settings management
@@ -327,17 +329,6 @@ async def google_auth_calendar(code: str = Query(...), Authorize: AuthJWT = Depe
     frontend_url = superagi.config.config.get_config("FRONTEND_URL", "http://localhost:3000")
     return RedirectResponse(frontend_url)
 
-@app.get('/oauth-twitter')
-async def twitter_oauth(oauth_token: str = Query(...),oauth_verifier: str = Query(...), Authorize: AuthJWT = Depends()):
-    token_uri = f'https://api.twitter.com/oauth/access_token?oauth_verifier={oauth_verifier}&oauth_token={oauth_token}'
-    conn = http_client.HTTPSConnection("api.twitter.com")
-    conn.request("POST", token_uri, "")
-    res = conn.getresponse()
-    response_data = res.read().decode('utf-8')
-    frontend_url = superagi.config.config.get_config("FRONTEND_URL", "http://localhost:3000")
-    redirect_url_success = f"{frontend_url}/twitter_creds/?{response_data}"
-    return RedirectResponse(url=redirect_url_success)
-
 @app.get('/github-login')
 def github_login():
     """GitHub login"""
@@ -427,41 +418,6 @@ def get_google_calendar_tool_configs(toolkit_id: int):
     return {
         "client_id": google_calendar_config.value
     }
-
-@app.get("/twitter/get_twitter_creds/toolkit_id/{toolkit_id}")
-def get_twitter_tool_configs(toolkit_id: int):
-    twitter_config_key = db.session.query(ToolConfig).filter(ToolConfig.toolkit_id == toolkit_id,ToolConfig.key == "TWITTER_API_KEY").first()
-    twitter_config_secret = db.session.query(ToolConfig).filter(ToolConfig.toolkit_id == toolkit_id,ToolConfig.key == "TWITTER_API_SECRET").first()
-    api_data =  {
-        "api_key": twitter_config_key.value,
-        "api_secret": twitter_config_secret.value
-    }
-    response = TwitterTokens().get_request_token(api_data)
-    return response
-
-@app.post("/twitter/send_twitter_creds/{twitter_creds}")
-def send_twitter_tool_configs(twitter_creds: str, Authorize: AuthJWT = Depends()):
-    Session = sessionmaker(bind=engine)
-    session = Session()
-    current_user = get_current_user()
-    user_id = current_user.id
-    credentials = json.loads(twitter_creds)
-    credentials["user_id"] = user_id
-    toolkit = session.query(Toolkit).filter(Toolkit.id == credentials["toolkit_id"]).first()
-    api_key = session.query(ToolConfig).filter(ToolConfig.key == "TWITTER_API_KEY", ToolConfig.toolkit_id == credentials["toolkit_id"]).first()
-    api_key_secret = session.query(ToolConfig).filter(ToolConfig.key == "TWITTER_API_SECRET", ToolConfig.toolkit_id == credentials["toolkit_id"]).first()
-    final_creds = {
-        "api_key": api_key.value,
-        "api_key_secret": api_key_secret.value,
-        "oauth_token": credentials["oauth_token"],
-        "oauth_token_secret": credentials["oauth_token_secret"]
-    }
-    tokens = OauthTokens().add_or_update(session, credentials["toolkit_id"], user_id, toolkit.organisation_id, "TWITTER_OAUTH_TOKENS", str(final_creds))
-    if tokens:
-        success = True
-    else:
-        success = False
-    return success
 
 @app.get("/validate-open-ai-key/{open_ai_key}")
 async def root(open_ai_key: str, Authorize: AuthJWT = Depends()):
